+++ pythautomata/tests/automata_definitions/other_automata.py
@@ -25,7 +25,7 @@ abcdeAlphabet = Alphabet(
 d = abcdeAlphabet['d']
 e = abcdeAlphabet['e']
 
-
+#TODO: DOCUMENT
 class OtherAutomata:
 
     @staticmethod
